breaking: taking element screenshot for iOS
- some other small fixes
